Move the channel about tab into it's own component
