+++ src/renderer/components/channel-about/channel-about.css
@@ -1,0 +1,42 @@
+.about {
+  background-color: var(--card-bg-color);
+  margin-top: 0;
+  padding: 10px;
+  position: relative;
+  z-index: 1;
+}
+
+.aboutInfo {
+  font-size: 17px;
+  white-space: pre-wrap;
+}
+
+.aboutTags {
+  display: flex;
+  flex-flow: row wrap;
+  gap: 5px 15px;
+  justify-content: center;
+  margin: 0;
+  padding: 0;
+}
+
+.aboutTag {
+  display: flex;
+  list-style: none;
+}
+
+.aboutTagLink {
+  background-color: var(--secondary-card-bg-color);
+  border-radius: 7px;
+  color: inherit;
+  padding: 7px;
+  text-decoration: none;
+}
+
+.aboutDetails {
+  text-align: left;
+}
+
+.aboutDetails th {
+  padding-right: 10px;
+}
