+++ src/renderer/views/Channel/Channel.css
@@ -122,50 +122,6 @@
   box-sizing: border-box;
 }
 
-.aboutTab {
-  background-color: var(--card-bg-color);
-  position: relative;
-  margin-top: 0px;
-  padding: 10px;
-  z-index: 1;
-}
-
-.aboutInfo {
-  font-family: 'Roboto', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
-  font-size: 17px;
-  white-space: pre-wrap;
-}
-
-.aboutTags {
-  display: flex;
-  flex-flow: row wrap;
-  gap: 5px 15px;
-  justify-content: center;
-  margin: 0;
-  padding: 0;
-}
-
-.aboutTag {
-  display: flex;
-  list-style: none;
-}
-
-.aboutTagLink {
-  background-color: var(--secondary-card-bg-color);
-  border-radius: 7px;
-  color: inherit;
-  padding: 7px;
-  text-decoration: none;
-}
-
-.aboutDetails {
-  text-align: left;
-}
-
-.aboutDetails th {
-  padding-right: 10px;
-}
-
 .channelSearch {
   margin-top: 10px;
   max-width: 250px;
